+++ src/components/pages/Order/Main/MainRightSide/Admin/Admin.jsx
@@ -1,8 +1,8 @@
 import { useContext } from "react"
 import styled from "styled-components"
 import OrderContext from "../../../../../../context/OrderContext"
-import AdminPanel from "../AdminPanel/AdminPanel"
-import AdminTabs from "./AdminTabs"
+import AdminPanel from "./AdminPanel/AdminPanel.jsx"
+import AdminTabs from "./AdminTabs.jsx"
 import { fadeInFromBottom } from "../../../../../../theme/animations"
 
 export default function Admin() {
