+++ src/components/pages/Order/Main/MainRightSide/Admin/AdminPanel/EditForm/EditInfoMessage.jsx
@@ -1,5 +1,5 @@
 import { styled } from "styled-components";
-import { theme } from "../../../../../../theme";
+import { theme } from "../../../../../../../../theme";
 
 export default function EditInfoMessage() {
   return (
