+++ src/components/pages/Order/Main/MainRightSide/Admin/tabsConfig.jsx
@@ -1,16 +1,16 @@
 import { AiOutlinePlus } from "react-icons/ai";
 import { MdModeEditOutline } from "react-icons/md";
-import AddForm from "../AdminPanel/AddForm.jsx";
-import EditForm from "../AdminPanel/EditForm";
-import HintMessage from "../AdminPanel/HintMessage.jsx";
+import AddForm from "./AdminPanel/AddForm/AddForm.jsx";
+import EditForm from "./AdminPanel/EditForm/EditForm";
+import HintMessage from "./AdminPanel/EditForm/HintMessage.jsx";
 
-export const getTabsConfig =  (hasAlreadyBeenClicked) => [
+export const getTabsConfig = (hasAlreadyBeenClicked) => [
   {
     index: "add",
     label: "Ajouter un produit",
     Icon: <AiOutlinePlus />,
     Content: <AddForm />,
-  },  
+  },
   {
     index: "edit",
     label: "Modifier un produit",
